Add Timer's Time Remaining to Digitial Watchface
Adds a live output of the timer's time remaining, along with
an hourGlass symbol to the left. Both of these are placed above
the current time and are the same color as the date as to not
draw attention away from the current time. The icon and the time
remaining are both set to hidden if the timer isn't running.