JustScott:digital_watchface_timer (#1)
* Add Timer's Time Remaining to Digitial Watchface

Adds a live output of the timer's time remaining, along with
an hourGlass symbol to the left. Both of these are placed above
the current time and are the same color as the date as to not
draw attention away from the current time. The icon and the time
remaining are both set to hidden if the timer isn't running.

* Use type uint8_t intead of type int for the timer's minute and second values

* Move the time remaining widget to StatusIcons

* Pass the timer as const

Doing this matches the other Controllers, which are also passed
as const to the StatusIcons.

---------

Co-authored-by: JustScott <development@justscott.me>