vkareh:global-widget-config (#2)
* Remove unused pointer to DisplayApp member variables (#2125)

In the screens that use `DisplayApp *app` and pass it to a child item,
or use the reference just in the constructor. Afterwards the `app`
member is not used. So remove it from the private member variables.

Completely remove `app` parameter from `SettingDisplay` constructor as
it is unused.

* MusicService: add missing includes for TickType_t and xTaskGetTickCount (#2130)

Add `FreeRTOS.h` include for the directly used data type `TickType_t` in the header
and the function `xTaskGetTickCount` from FreeRTOS's `task.h`

* settings: Add global widget selection

Instead of each watch face implementing their own settings for which
widgets to display, we can have a global selection of widgets. All
watch faces can then determine whether it is enabled and so display it
in whichever way makes sense for that face.

Current widgets supported are heart rate, step counter, and weather.

* WatchFaceAnalog: Add configurable widgets

* WatchFaceAnalog: Add weather widget

---------

Co-authored-by: NeroBurner <pyro4hell@gmail.com>
Co-authored-by: Victor Kareh <vkareh@redhat.com>